// App/src/config/constants/types.ts
export interface Address {
  97?: string
  56?: string
  25?: string
  80001?: string
}

export interface SerializedToken {
  chainId: number
  address: string
  decimals: number
  symbol?: string
  name?: string
  projectLink?: string
}
// ...
interface NftFarmGangverseConfigBaseProps {
  pid: number
  lpSymbol: string
  nftAddresses: Address
  contractAddresses?: Address
  multiplier?: string
  tokenPerBlock?: string
  participantThreshold?: number
  isFinished?: boolean
  performanceFee?: string
  supportedCollectionPids?:number[]
  dual?: {
    rewardPerBlock: number
    earnLabel: string
    endBlock: number
  }
}

// New type for NFT Stake config
export interface SerializedNftFarmConfig extends NftFarmConfigBaseProps {
  earningToken?: SerializedToken
  projectLink?: string
}

export interface SerializedNftFarmGangverseConfig extends NftFarmGangverseConfigBaseProps {
  earningToken?: SerializedToken
  projectLink?: string
}